src/requirements.txt
beautifulsoup4==4.12.3
Flask==2.2.2
Flask_Login==0.6.2
flask_sqlalchemy==3.1.1
lxml==4.9.3
matplotlib==3.8.0
memory_profiler==0.58.0
neo4j==5.14.1
NER==0.1
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
pandas==2.0.3
pydantic==1.10.13
pytorch_crf==0.7.2
pyvis==0.3.1
PyYAML==6.0.1
PyYAML==5.3.1
PyYAML==6.0.1
redis==5.0.1
Requests==2.31.0
scikit_learn==1.3.0
scipy==1.12.0
spacy==3.5.3
torch==1.13.0
torchtext==0.14.0
tqdm==4.66.1
transformers==4.30.2
Werkzeug==2.2.3
Whoosh==2.7.4
